Move button to detailed stream stats folder
